requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==3.0.1
email-validator==2.1.0.post1
cryptography==42.0.2 